Initial Commit
Here's everything.
I don't like it, but I believe that LWJGL and JBullet are getting
uploaded as well. It doesn't really matter as they are both okay with
this matter (LWJGL is under a BSD license and JBullet is under a ZLIB
license).
I am not an author of LWJGL or JBullet.