Replaced Hashtables with HashMaps
I thought that Hashtables and HashMaps were the same, but it turns out
that Hashtables are based on the old Dictionary class, which is now
deprecated. HashMaps are newer, but for whatever reason aren't
synchronized. I'll cross that bridge when I come to it (which 'll
probably be never, since this should be single-threaded).